x86/sgx: fix a NULL pointer
Under heavy load, the SGX EPC reclaimers (ksgxd or future EPC cgroup
worker) may reclaim SECS EPC page for an enclave and set
encl->secs.epc_page to NULL.  And the SECS EPC page is required for EAUG
in #PF handler and is used without checking and loading.

Fix this by checking if SECS is loaded beofre EAUG and load it if it was
reclaimed.

Signed-off-by: Haitao Huang <haitao.huang@linux.intel.com>

V3:
- Added